Remove dirty build-graphite/ before building
The build process should remove build-graphite before building. In the scenario where a clean copy of the repo is used, no harm done. In the scenario where the build broke in the middle or was canceled. the next build would see the build-graphite/ directory and just create a broken docker image that's halfway built.